+++ modalities/files/doktok_modalities_files/__init__.py
@@ -2,7 +2,14 @@
 
 from doktok_modalities_files.extractors import DirectTextExtractor, PyMuPdfTextExtractor
 from doktok_modalities_files.mime import LibmagicMimeDetector
+from doktok_modalities_files.render import PyMuPdfRenderer, SearchablePdfBuilder
 
 __version__ = "0.0.0"
 
-__all__ = ["DirectTextExtractor", "LibmagicMimeDetector", "PyMuPdfTextExtractor"]
+__all__ = [
+    "DirectTextExtractor",
+    "LibmagicMimeDetector",
+    "PyMuPdfRenderer",
+    "PyMuPdfTextExtractor",
+    "SearchablePdfBuilder",
+]